use size_t as argumets to gp_file_slurp
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@9764 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ libgphoto2/gphoto2-file.c
@@ -238,7 +238,7 @@ gp_file_append (CameraFile *file, const char *data,
  **/
 int
 gp_file_slurp (CameraFile *file, char *data, 
-	unsigned long size, unsigned long *readlen
+	size_t size, size_t *readlen
 ) {
 	CHECK_NULL (file);
